refactor #56: redis에 저장된 조회수를 응답값에 넣도록 변경

+++ src/main/java/com/codesquad/secondhand/domain/item/Item.java
@@ -73,7 +73,7 @@ public class Item extends BaseTimeEntity {
 	private String title;
 	private String content;
 	private Integer price;
-	private int views;
+	private Long views;
 	private boolean isDeleted;
 
 	public Item(User user, Category category, Region region, Status status,
@@ -87,7 +87,7 @@ public class Item extends BaseTimeEntity {
 		this.title = title;
 		this.content = content;
 		this.price = price;
-		this.views = 0;
+		this.views = 0L;
 		this.isDeleted = false;
 	}
